Refactor | Moved the values to the newly created data stores

+++ app/src/main/java/com/example/blemeter/config/extenstions/UByteExtensions.kt
@@ -8,7 +8,7 @@ import java.nio.ByteOrder
 fun String.fromHexToUByteArray(): UByteArray {
     check(length % 2 == 0) { "Must have an even length" }
 
-    return chunked(2).map { it.toInt(16).toUByte() }.toUByteArray()
+    return chunked(2).map { it.toUByte(16) }.toUByteArray()
 }
 
 @OptIn(ExperimentalUnsignedTypes::class)
